+++ lib/auditor_qt.rb
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 # encoding: utf-8
 
-# File: auditer_qt.rb
+# File: auditor_qt.rb
 # Created: 18/06/12
 #
 # (c) Michel Demazure <michel@demazure.com>
@@ -11,8 +11,8 @@ require 'j2r/jaccess'
 require 'j2r/core'
 
 J2R.for_user do
-  J2R.logger.info('start auditer')
-  require_relative('j2r/Qt/auditer_main.rb')
-  J2R::GuiQt::AuditerMain.run
-  J2R.logger.info('stop auditer')
+  J2R.logger.info('start auditor')
+  require_relative('j2r/Qt/auditor_main.rb')
+  J2R::GuiQt::AuditorMain.run
+  J2R.logger.info('stop auditor')
 end