+++ lib/j2r/Qt/auditor_elements.rb
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 # encoding: utf-8
 
-# File: auditer_elements.rb
+# File: auditor_elements.rb
 # Created 27/06/12
 #
 # (c) Michel Demazure <michel@demazure.com>
@@ -13,7 +13,7 @@ require_relative 'frames/dialogs.rb'
 require_relative 'frames/pretty_frame.rb'
 require_relative 'frames/pretty_combo.rb'
 require_relative 'frames/mailing_frame.rb'
-require_relative 'frames/audit_frame.rb'
+require_relative 'frames/tiers_audit_frame.rb'
 require_relative 'frames/jacinthe_main.rb'
 
-require_relative 'frames/auditer_central_widget.rb'
+require_relative 'frames/auditor_central_widget.rb'